+++ src/pages/AddContact.jsx
@@ -1,7 +1,10 @@
-import React from 'react'
+import {useContext} from 'react'
 import ContactForm from '../Components/contacts/ContactForm'
+import { ContactContext } from '../context/Contact.context'
 
-function Addcontact({addContact}){
+function Addcontact(){
+  const {addContact} = useContext(ContactContext)
+  
   return (
     <>
       <ContactForm addContact={addContact} />
