+++ src/pages/ContactDetails.jsx
@@ -1,12 +1,14 @@
-import { useEffect, useState } from 'react'
+import { useEffect, useState, useContext } from 'react'
 import { useParams, Link, useNavigate } from 'react-router-dom'
 import { Container, Card, Button, ListGroup, } from 'react-bootstrap'
 import { BsFillTrashFill} from 'react-icons/bs'
 import {FaRegEdit} from 'react-icons/fa'
 import { toast } from 'react-toastify'
 import {format} from 'date-fns'
+import { ContactContext } from '../context/Contact.context'
 
-function ContactDetails({contacts, deleteContact}) {
+function ContactDetails() {
+    const {contacts, deleteContact} = useContext(ContactContext)
     const [newCont, setNewCont] = useState({})
     const params = useParams()
     const {id} = params
